Refactor transaction helpers for better organization
