+++ tsconfig.json
@@ -21,7 +21,7 @@
     "paths": {
       "@helpers/client": ["./helpers/client.ts"],
       "@helpers/utils": ["./helpers/utils.ts"],
-      "@helpers/transaction": ["./helpers/transaction.ts"]
+      "@helpers/usdc": ["helpers/usdc.ts"]
     }
   },
   "include": ["**/*.ts"]
